requirements/backend.txt
# ------------------------
# Core AI / LLM
# ------------------------
qdrant-client>=1.9.0,<2.0.0
langchain==0.3.27
langchain-core>=0.3.70,<0.3.80
langchain-community==0.3.27
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.9
langchain-huggingface>=0.3.0
langsmith==0.3.45
sentence-transformers==2.7.0
huggingface-hub==0.33.4
openai>=1.104.2,<2.0.0
transformers==4.53.2
tokenizers==0.21.2
safetensors==0.5.3
tiktoken==0.9.0
torch>=2.0.0,<3.0.0
# ------------------------
# Web server / MCP
# ------------------------
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
starlette>=0.38.0,<0.39.0
sse-starlette>=1.6.5,<2.0.0
mcp[cli]>=1.0.0
# ------------------------
# Authentication / Config
# ------------------------
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
# ------------------------
# Pydantic v2 stack
# ------------------------
pydantic>=2.11.0,<2.12.0
pydantic-settings==2.10.1
pydantic-core==2.33.2
# ------------------------
# Networking / HTTP
# ------------------------
httpx>=0.27.0,<0.29.0
httpcore>=1.0.0,<2.0.0
httpx-sse==0.4.1
aiohttp==3.12.14
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
requests>=2.31.0,<3.0.0
urllib3>=1.26.15,<3.0.0
# ------------------------
# Data Science / ML Utils
# ------------------------
numpy==1.26.4
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
joblib==1.2.0
threadpoolctl==3.1.0
# ------------------------
# Utilities
# ------------------------
tenacity==9.1.2
tqdm==4.67.1
filelock==3.18.0
fsspec==2025.5.1
regex==2024.11.6
orjson==3.10.18
jsonpatch==1.33
jsonpointer==3.0.0
PyYAML==6.0.2
lxml==5.2.1
loguru==0.7.3
# ------------------------
# DB / ORM
# ------------------------
SQLAlchemy>=2.0.30,<2.0.36
supabase>=2.18.0,<3.0.0
cachetools>=5.3.3,<6.0.0